pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "natural2sql"
version = "1.0.0"
description = "A multi-turn SQLite environment simulating an e-commerce database."
requires-python = ">=3.11"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic"
]

[project.scripts]
server = "server.app:main"